move avatar size logic to the new funciton getAvatarSize
